Changed so that the drainer thread grabs all of the requests on the queue instead of grabbing just one (reduces contention for the queue).  This can cause performance to be choppy on uniprocessors; it might not be the right thing to do everywhere.
Fixed a bug in managing the soft-references to cached data.  Plugs a small
memory leak.

Added a progressive throttle to the diskUpdateQueue: as the queue gets longer,
threads start to block for longer and longer periods of time when they l
enqueue something, giving the queue more time to drain.  A very simple
mechanism and the constants are all guesses...  Should be tuned (or tune
itself) to new hardware and operating environments.

Fixed bug in close(): all the connections were being left open.  Also
being more aggressive about freeing the contents of cached tables (although
this should be taken care of automatically).

-DJE