Walkmesh should not be consulted when no motion has occured.
Exception from walkmesh should be treated as a cheat-attempt, as it
ususally means someone off the map (but a remaining bug in WalkMeshMap
throws NPE possibly on tile boundaries).

When a cheat is detected, send player back to the start of this walk segment,
as opposed to the hit-location (which requires computation of avatar width).