Major bug fixes to transactions. 1. A transaction upgrading PEEK to GET would incorrectly steal the lock from another transaction that had done an intervening GET.  Both transactions would think they had the lock, and both would later release it.  Combined with bug 2 below, this was bad news.
This was primarily due to a subtle caching issue at the DataStoreTransactionImpl layer; the solution was to purge the cache entry before reading the object's lock header.

2. In cases where a stale lock was grabbed by another transaction, the original transaction would still treat the lock as its own and free it during commit or abort.  This often led to further, cascading problems: other transactions then obtained the lock thinking it was free, only to have it freed out from under them, but then they freed it anyway, etc.

Transactions now check each lock during commit and abort to make sure they're still the owner.  If during commit they find they lost a lock, they abort instead.  If they find it during abort, they skip that lock (don't free it).  This also required the cache-flushing technique above so the change in ownership would be detected.

3. During a GET, transaction A might find that the lock has just become free, but discover a more senior transaction B that is blocked waiting on it which hasn't woken up to acquire it yet.  In that case, A is supposed to assign ownership to B and try again.  This was broken if A was doing an ATTEMPT (non-blocking GET).

In that case, A would erroneously add itself to the wait-set for the lock before bowing out, even though it was not going to wait.  Once B had committed, other transactions would see A as the senior waiter and assign the lock to A.  Since A was oblivious to the lock, it was not freed when transaction A ended, and all other transactions usually had to wait until the lock became stale to grab it (in the EA2 stack, that's about 10 seconds).