Patch to frozen trunk approved by Jim. Reviewed by Ann. Merges [2127] from branch jmegq.  Details:
ClientSessionImpl:
In sendProtocolMessage, reduce the log level in the case where the session is disconnected; this may happen due to the cached channel state and the current message at level WARNING is alarming.
Remove the hexdump of the message in the two non-FINEST log cases; it is too big and too slow.

TestChannelServiceImpl:
In checkMembership, handle the case where sessionService#getClientSession returns null.

TestChannelServiceImpl and TestClientSessionServiceImpl:
Add a 100ms pause between each test so that a lock notification triggered asynchronously by the previous test cannot interfere with the next test.